Tweaks/refactor for tfrecord predictions
