+++ make_predictions.py
@@ -8,7 +8,7 @@ import tensorflow as tf
 from zoobot import label_metadata
 from zoobot.data_utils import image_datasets
 from zoobot.estimators import define_model, preprocess
-from zoobot.predictions import predict_on_tfrecords, predict_on_images
+from zoobot.predictions import predict_on_tfrecords, predict_on_dataset
 
 
 if __name__ == '__main__':
@@ -27,7 +27,7 @@ if __name__ == '__main__':
     file_format = 'png'
 
     # utility function to easily list the images in a folder.
-    unordered_image_paths = predict_on_images.paths_in_folder('data/example_images/basic', file_format=file_format, recursive=False)
+    unordered_image_paths = predict_on_dataset.paths_in_folder('data/example_images/basic', file_format=file_format, recursive=False)
 
     ## or maybe you already have a list from a catalog?
     # unordered_image_paths = df['paths']
@@ -113,4 +113,4 @@ if __name__ == '__main__':
     # save_loc = 'data/results/make_predictions_example.csv'
     save_loc = 'data/results/make_predictions_example.hdf5'
     n_samples = 5
-    predict_on_images.predict(image_ds, model, n_samples, label_cols, save_loc)
+    predict_on_dataset.predict(image_ds, model, n_samples, label_cols, save_loc)
